add show slider in frontend page

+++ app/Http/Controllers/Admin/SliderController.php
@@ -132,7 +132,9 @@ class SliderController extends Controller
     public function destroy($id)
     {
         $slider = Slider::find($id);
-        unlink('uploads/slider/'. $slider->image);
+        if (file_exists('uploads/slider/'. $slider->image)) {
+            unlink('uploads/slider/'. $slider->image);
+        }
         $slider->delete();
         return redirect()->back()->with('successMsg', 'Slider Successfuly Deleted');
     }
